move UniqueWordCounter to util folder added private to all enum constructors as required
