+++ src/main/java/com/solvd/enums/JobPosition.java
@@ -10,7 +10,7 @@ public enum JobPosition {
 
     private final String jobPosition;
 
-    JobPosition(String jobPosition){
+    private JobPosition(String jobPosition){
 
         this.jobPosition = jobPosition;
     }
