+++ src/main/java/com/solvd/enums/Schedule.java
@@ -12,7 +12,7 @@ public enum Schedule {
 
     private final String dayName;
 
-    Schedule(String dayName) {
+    private Schedule(String dayName) {
         this.dayName = dayName;
     }
